Seguridad y restriccion de Usuarios

+++ resources/views/entrenadores/front_perfil_entrenador.blade.php
@@ -47,30 +47,35 @@
       <ul class="d-flex align-items-end flex-column fixed-bottom text-white">
         <li id="registrar-entrenador" class="p-2">
           <a type="button" class="btn btn-warning" onclick="editEntre();">Editar datos</a>
-          <a type="button" class="btn btn-danger" data-toggle="modal" data-target="#deleteEntrenador">Eliminar entrenador</a>
+
+           @if(Auth::user()->idtipoUsuario == 1)
+              <a type="button" class="btn btn-danger" data-toggle="modal" data-target="#deleteEntrenador">Eliminar entrenador</a>
+           @endif
+
         </li>
       </ul>
 
-
-      <div class="modal fade" id="deleteEntrenador" tabindex="-1" role="dialog" aria-labelledby="deleteEntrenador" aria-hidden="true">
-        <div class="modal-dialog" role="document">
-          <div class="modal-content">
-            <div class="modal-header">
-              <h5 class="modal-title" id="deleteEntrenador">Eliminar entrenador</h5>
-              <button type="button" class="close" data-dismiss="modal" aria-label="Close">
-                <span aria-hidden="true">&times;</span>
-              </button>
-            </div>
-            <div class="modal-body">
-              <p>Esta acción no se podrá deshacer, los competidores con este entrenador pasarán a no tener ninguno</p>
-            </div>
-            <div class="modal-footer">
-              <button type="button" class="btn btn-secondary" data-dismiss="modal">Cancelar</button>
-              <button type="button" class="btn btn-danger" onclick="deleteEntrenador()">Eliminar</button>
+      @if(Auth::user()->idtipoUsuario == 1)
+          <div class="modal fade" id="deleteEntrenador" tabindex="-1" role="dialog" aria-labelledby="deleteEntrenador" aria-hidden="true">
+            <div class="modal-dialog" role="document">
+              <div class="modal-content">
+                <div class="modal-header">
+                  <h5 class="modal-title" id="deleteEntrenador">Eliminar entrenador</h5>
+                  <button type="button" class="close" data-dismiss="modal" aria-label="Close">
+                    <span aria-hidden="true">&times;</span>
+                  </button>
+                </div>
+                <div class="modal-body">
+                  <p>Esta acción no se podrá deshacer, los competidores con este entrenador pasarán a no tener ninguno</p>
+                </div>
+                <div class="modal-footer">
+                  <button type="button" class="btn btn-secondary" data-dismiss="modal">Cancelar</button>
+                  <button type="button" class="btn btn-danger" onclick="deleteEntrenador()">Eliminar</button>
+                </div>
+              </div>
             </div>
           </div>
-        </div>
-      </div>
+      @endif
     @endif
 @else
   <h4>No tienes permisos para realizar esto... </h4>
